[Next Gen] refactor(codegen): split v-for item prop object generation into item_props
