+++ crates/vize_atelier_core/src/codegen/v_for.rs
@@ -5,6 +5,7 @@
 
 mod generate;
 pub(crate) mod helpers;
+mod item_props;
 mod item_props_merge;
 
 use crate::steps::v_memo::{get_memo_exp, has_v_memo};
